pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codingcrew"
version = "0.2.0"
description = "Autonome Coding-Crew mit zentraler YAML-Config. Kein LiteLLM — direkte Provider-Calls."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
cloud = [
    "anthropic>=0.20",
    "openai>=1.0",
]
api = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
codingcrew = "src.orchestrator:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
